# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Obsura API",
    "summary": "Review-first sanitization workflows, studio assets, and job orchestration for Obsura.",
    "description": "Obsura API is the review-first workflow and orchestration layer for Obsura.\n\nThe API supports:\n- persistent studio assets for custom detection and reusable configurations\n- reviewable jobs and run history\n- text-first and screenshot-first sanitization workflows\n- transformation behavior beyond plain placeholder replacement\n- image-region protection flows for blur, pixelation, masks, and overlays\n\nThis OpenAPI document is intended for direct import into API clients such as Postman.",
    "contact": {
      "name": "Obsura",
      "url": "https://github.com/obsura"
    },
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List persisted jobs with findings and generated outputs.",
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_JobRead__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "description": "Fetch one persisted job by identifier.",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_JobRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/review": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Review Job",
        "description": "Apply review decisions to findings within a persisted job.",
        "operationId": "review_job_api_v1_jobs__job_id__review_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_JobRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workflows/images/analyze": {
      "post": {
        "tags": [
          "image-workflows"
        ],
        "summary": "Analyze Image",
        "description": "Analyze an uploaded image or screenshot into reviewable regions.",
        "operationId": "analyze_image_api_v1_workflows_images_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_image_api_v1_workflows_images_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ImageWorkflowResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/images/transform": {
      "post": {
        "tags": [
          "image-workflows"
        ],
        "summary": "Transform Image",
        "description": "Apply image-region transformations and persist the generated output.",
        "operationId": "transform_image_api_v1_workflows_images_transform_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transform_image_api_v1_workflows_images_transform_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ImageWorkflowResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/images/transform-job": {
      "post": {
        "tags": [
          "image-workflows"
        ],
        "summary": "Transform Image Job",
        "description": "Transform a persisted image or screenshot job after review decisions are applied.",
        "operationId": "transform_image_job_api_v1_workflows_images_transform_job_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageJobTransformRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ImageWorkflowResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "description": "Unified error payload."
      },
...
      "ApiResponse_ImageWorkflowResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageWorkflowResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ApiResponse[ImageWorkflowResponse]"
      },
      "ApiResponse_JobRead_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ApiResponse[JobRead]"
      },
...
      "ApiResponse_list_JobRead__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JobRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ApiResponse[list[JobRead]]"
      },
...
      "Body_analyze_image_api_v1_workflows_images_analyze_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "manifest_json": {
            "type": "string",
            "title": "Manifest Json",
            "default": "{}"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_image_api_v1_workflows_images_analyze_post"
      },
      "Body_transform_image_api_v1_workflows_images_transform_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "manifest_json": {
            "type": "string",
            "title": "Manifest Json",
            "default": "{}"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_transform_image_api_v1_workflows_images_transform_post"
      },
      "BoundingBox": {
        "properties": {
          "x": {
            "type": "integer",
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BoundingBox",
        "description": "A region on an image or screenshot."
      },
...
      "ContentType": {
        "type": "string",
        "enum": [
          "text",
          "structured_text",
          "screenshot",
          "image"
        ],
        "title": "ContentType"
      },
...
      "FindingKind": {
        "type": "string",
        "enum": [
          "text_span",
          "image_region",
          "face_region",
          "face_subregion"
        ],
        "title": "FindingKind"
      },
...
      "FindingRecord": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "source": {
            "$ref": "#/components/schemas/FindingSource"
          },
          "kind": {
            "$ref": "#/components/schemas/FindingKind"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Name"
          },
          "start_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Index"
          },
          "end_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Index"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundingBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "matched_text_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Text Preview"
          },
          "matched_text_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Text Hash"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1.0
          },
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision",
            "default": "pending"
          },
          "transformation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformationRule"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            },
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "source",
          "kind",
          "entity_type"
        ],
        "title": "FindingRecord",
        "description": "Reviewable finding for text or image workflows."
      },
      "FindingSource": {
        "type": "string",
        "enum": [
          "built_in",
          "custom",
          "manual",
          "face",
          "ocr"
        ],
        "title": "FindingSource"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageFindingOverride": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "transformation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformationRule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "finding_id"
        ],
        "title": "ImageFindingOverride",
        "description": "Override a stored image finding before export."
      },
      "ImageJobTransformRequest": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "finding_overrides": {
            "items": {
              "$ref": "#/components/schemas/ImageFindingOverride"
            },
            "type": "array",
            "title": "Finding Overrides"
          },
          "include_pending": {
            "type": "boolean",
            "title": "Include Pending",
            "default": false
          },
          "persist_output": {
            "type": "boolean",
            "title": "Persist Output",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "ImageJobTransformRequest",
        "description": "Request to transform a persisted image job after review."
      },
      "ImageWorkflowResponse": {
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingRecord"
            },
            "type": "array",
            "title": "Findings"
          },
          "stored_input_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored Input Path"
          },
          "stored_output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored Output Path"
          },
          "media_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Url"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "findings",
          "summary"
        ],
        "title": "ImageWorkflowResponse",
        "description": "Response for image analysis or transformation."
      },
      "JobOutputRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "output_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Text"
          },
          "output_file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output File Path"
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            },
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "job_id",
          "content_type"
        ],
        "title": "JobOutputRecord",
        "description": "A stored output generated from a job."
      },
      "JobRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text"
          },
          "source_file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File Path"
          },
          "pattern_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pattern Ids"
          },
          "custom_entity_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Custom Entity Ids"
          },
          "configuration_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Configuration Ids"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingRecord"
            },
            "type": "array",
            "title": "Findings"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/JobOutputRecord"
            },
            "type": "array",
            "title": "Outputs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "status",
          "content_type"
        ],
        "title": "JobRead",
        "description": "Full stored job representation."
      },
      "JobReviewDecisionInput": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision"
          },
          "transformation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformationRule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "finding_id",
          "decision"
        ],
        "title": "JobReviewDecisionInput",
        "description": "Review decision for a single finding."
      },
      "JobReviewRequest": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/JobReviewDecisionInput"
            },
            "type": "array",
            "title": "Decisions"
          }
        },
        "type": "object",
        "title": "JobReviewRequest",
        "description": "Batch review update for stored findings."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "analyzed",
          "reviewed",
          "transformed"
        ],
        "title": "JobStatus"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta",
        "description": "Pagination metadata returned alongside paged results."
      },
...
      "ReviewDecision": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "ReviewDecision"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}