# src/obsura_api/api/routes/image_workflows.py
"""Routes for image-region workflows."""

from __future__ import annotations

from typing import Annotated

from fastapi import APIRouter, Depends, File, Form, UploadFile
from sqlalchemy.orm import Session

from obsura_api.api.dependencies import get_container, get_db_session, get_settings
from obsura_api.api.responses import success_response
from obsura_api.core.container import AppContainer
from obsura_api.core.settings import Settings
from obsura_api.domain.common import ApiResponse
from obsura_api.domain.workflows import (
    ImageJobTransformRequest,
    ImageWorkflowManifest,
    ImageWorkflowResponse,
)
from obsura_api.services.image_workflows import ImageWorkflowService

router = APIRouter(prefix="/workflows/images", tags=["image-workflows"])
SessionDep = Annotated[Session, Depends(get_db_session)]
SettingsDep = Annotated[Settings, Depends(get_settings)]
ContainerDep = Annotated[AppContainer, Depends(get_container)]


@router.post("/analyze", response_model=ApiResponse[ImageWorkflowResponse])
async def analyze_image(
    session: SessionDep,
    settings: SettingsDep,
    container: ContainerDep,
    file: UploadFile = File(...),
    manifest_json: str = Form("{}"),
) -> ApiResponse[ImageWorkflowResponse]:
    """Analyze an uploaded image or screenshot into reviewable regions."""

    manifest = ImageWorkflowManifest.model_validate_json(manifest_json)
    service = ImageWorkflowService(
        session=session,
        settings=settings,
        storage=container.storage,
        face_detector=container.face_detector,
    )
    return success_response(
        service.analyze(
            file_bytes=await file.read(),
            filename=file.filename or "upload.bin",
            manifest=manifest,
        ),
    )


@router.post("/transform", response_model=ApiResponse[ImageWorkflowResponse])
async def transform_image(
    session: SessionDep,
    settings: SettingsDep,
    container: ContainerDep,
    file: UploadFile = File(...),
    manifest_json: str = Form("{}"),
) -> ApiResponse[ImageWorkflowResponse]:
    """Apply image-region transformations and persist the generated output."""

    manifest = ImageWorkflowManifest.model_validate_json(manifest_json)
    service = ImageWorkflowService(
        session=session,
        settings=settings,
        storage=container.storage,
        face_detector=container.face_detector,
    )
    return success_response(
        service.transform(
            file_bytes=await file.read(),
            filename=file.filename or "upload.bin",
            manifest=manifest,
        ),
    )


@router.post("/transform-job", response_model=ApiResponse[ImageWorkflowResponse])
def transform_image_job(
    payload: ImageJobTransformRequest,
    session: SessionDep,
    settings: SettingsDep,
    container: ContainerDep,
) -> ApiResponse[ImageWorkflowResponse]:
    """Transform a persisted image or screenshot job after review decisions are applied."""

    service = ImageWorkflowService(
        session=session,
        settings=settings,
        storage=container.storage,
        face_detector=container.face_detector,
    )
    return success_response(service.transform_job(payload))